Reorganize terrain generation files [Re: nmkataoka/adv-life#226]
