+++ src/1-game-code/World/TerrainGen/lessThan.ts
@@ -1,5 +1,5 @@
 import { Vector2 } from '8-helpers/math/Vector2';
-import { VorEdge } from './VorEdge';
+import { VorEdge } from './Voronoi/VorEdge';
 
 /** Sorts by x first, then y. Useful if a consistent order is desired for a set map-related points. */
 export function lessThan(a: Vector2, b: Vector2): number {
