+++ src/1-game-code/World/TerrainGen/tectonicGeneration/index.ts
@@ -1,0 +1,1 @@
+export { generateTectonics } from './generateTectonics';
