+++ src/1-game-code/World/TerrainGen/tectonicRasterization/fillInHoles.test.ts
@@ -1,4 +1,4 @@
-import { DataLayer } from '../DataLayer';
+import { DataLayer } from '../../DataLayer';
 import { findMostCommonElevationOnHoleBorder } from './fillInHoles';
 
 describe('fillInHoles', () => {
