+++ src/1-game-code/World/TerrainGen/tectonicRasterization/floodfillFromFault.ts
@@ -2,8 +2,8 @@ import { initializeArrayWithValue, shuffle } from '8-helpers/ArrayExtensions';
 import { RingQueue } from '8-helpers/containers/RingQueue';
 import { Vector2 } from '8-helpers/math';
 import { add, dist, multiply } from '8-helpers/math/Vector2';
-import { DataLayer } from '../DataLayer';
-import { Fault } from './Fault';
+import { DataLayer } from '../../DataLayer';
+import { Fault } from '../Fault';
 import { simpleBresenham } from './simpleBresenham';
 
 type TilePropagator = {
