+++ src/1-game-code/World/TerrainGen/tectonicRasterization/index.ts
@@ -1,0 +1,4 @@
+export { rasterizeTectonics } from './rasterizeTecPlates';
+export { shapeCoasts } from './shapeCoasts';
+export { fillInHoles } from './fillInHoles';
+export { rasterizeFaults } from './rasterizeFaults';
