+++ src/1-game-code/World/TerrainGen/tectonicRasterization/shapeCoasts.ts
@@ -1,10 +1,10 @@
 import { initializeArrayWithValue } from '8-helpers/ArrayExtensions';
 import { RingQueue } from '8-helpers/containers/RingQueue';
 import { toVec2i, Vector2 } from '8-helpers/math/Vector2';
-import { DataLayer } from '../DataLayer';
-import { Fault } from './Fault';
+import { DataLayer } from '../../DataLayer';
+import { Fault } from '../Fault';
 import { floodfillFromFault } from './floodfillFromFault';
-import { TecPlate } from './TecPlate';
+import { TecPlate } from '../TecPlate';
 
 export function shapeCoasts(elevLayer: DataLayer, faults: Fault[], tecPlates: TecPlate[]): void {
   const { height, width } = elevLayer;
